Revert "GO1P-25335 Add abstraction layer for cloud storage"
