+++ .travis.yml
@@ -1,6 +1,5 @@
 language: php
 php:
 - 7.2
-- 7.3
 before_script: composer install --prefer-source
 script: phpunit
