+++ composer.json
@@ -11,8 +11,6 @@
   "require": {
     "elasticsearch/elasticsearch": "^5.0",
     "rappasoft/laravel-helpers": "dev-master",
-    "go1/util": "dev-master",
-    "mrubiosan/flysystem-url": "^1.0",
-    "league/flysystem-aws-s3-v3": "^1.0"
+    "go1/util": "dev-master"
   }
 }
